trying to start rest template

+++ src/main/java/lits/jp/hotel/management/models/PlacesRequest.java
@@ -1,12 +1,13 @@
 package lits.jp.hotel.management.models;
 
-import lombok.Data;
+import lombok.*;
 
 @Data
+
 public class PlacesRequest {
   String provider;
   String type;
   Integer radius;
-  Long lat;
-  Long lng;
+  Double lat;
+  Double lng;
 }
